1. Add activity - user

+++ src/main/java/com/cobble/takeaway/pojo/RelUserRolePOJO.java
@@ -1,26 +1,26 @@
 package com.cobble.takeaway.pojo;
 
 public class RelUserRolePOJO extends BasePOJO {
-    private Integer relUserId;
-    private Integer userId;
-    private Integer roleId;
+    private Long relUserId;
+    private Long userId;
+    private Long roleId;
     
-	public Integer getRelUserId() {
+	public Long getRelUserId() {
 		return relUserId;
 	}
-	public void setRelUserId(Integer relUserId) {
+	public void setRelUserId(Long relUserId) {
 		this.relUserId = relUserId;
 	}
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
-	public Integer getRoleId() {
+	public Long getRoleId() {
 		return roleId;
 	}
-	public void setRoleId(Integer roleId) {
+	public void setRoleId(Long roleId) {
 		this.roleId = roleId;
 	}
